common/xmalloc_tlsf: change bitmaps to unsigned long instead of u32
Atomic operations are used on these bitmaps and atomics take a performance
hit on some architectures if they are not aligned to the architecture's
integer width (64-bit on x86_64, for example).  On other architectures,
misaligned atomics are simply not supported (RISC-V).

This commit changes these bitmaps to be "unsigned long" in order to
cause these fields to be automatically aligned to the integer width of
the architecture when compiled with strict alignment or on compiler
targets that default to strict alignment.

Signed-off-by: Bobby Eshleman <bobby.eshleman@gmail.com>